+++ pkg/common/cmsql/cmsql.go
@@ -1,4 +1,4 @@
-package sql
+package cmsql
 
 import (
 	"fmt"
